feat: edit member part and comment
