+++ .gitignore
@@ -33,3 +33,5 @@ Thumbs.db
 project
 classpath
 settings/
+
+config/
