--- perfectJournal/src/main/java/com/pj/journal/model/member/LoginDao.java
@@ -1,13 +1,0 @@
-package com.pj.journal.model.member;
-
-public interface LoginDao {
-
-	LoginVo selectOneMember(int pk);
-
-	void insertOneMember(LoginVo bean);
-
-	int updateOneMember(LoginVo bean);
-
-	int deleteOneMember(int pk);
-
-}
