+++ perfectJournal/src/main/resources/application.properties
@@ -1,1 +1,7 @@
 spring.application.name=perfectJournal
+
+mybatis.config-location=classpath:mybatis-config.xml
+
+mybatis.mapper-locations=classpath:/mappers/*-mapper.xml
+
+
