+++ perfectJournal/src/main/resources/mappers/Member-mapper.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.pj.journal.model.member.MemberDao">
+
+	<select id="loginOneMember" parameterType="map" resultType="MemberBean">
+		SELECT * FROM users
+		WHERE user = #{user} AND password = #{password}
+	</select>
+
+	<select id="findMemberId" parameterType="map" resultType="String">
+		SELECT user FROM users
+		WHERE email = #{email} AND answer = #{answer} AND question = #{question}
+	</select>
+
+	<select id="findMemberPw" parameterType="map" resultType="MemberBean">
+		SELECT user FROM users
+		WHERE email = #{email}
+		AND answer = #{answer}
+		AND user = #{user}
+		AND question = #{question}
+	</select>
+
+	<update id="changeMemberPw" parameterType="map">
+		UPDATE users SET password = #{password}
+		WHERE user = #{user}
+	</update>
+
+	<insert id="insertOneMember" parameterType="MemberBean">
+		INSERT INTO users (email, user, password, nickname, created_at,
+		question, answer)
+		VALUES (#{email}, #{user}, #{password}, #{nickname}, now(), #{question},
+		#{answer})
+	</insert>
+
+</mapper>
